+++ operator-common/src/test/java/io/strimzi/operator/common/operator/resource/AbstractCustomResourceOperatorIT.java
@@ -124,7 +124,7 @@ public abstract class AbstractCustomResourceOperatorIT<C extends KubernetesClien
         PlatformFeaturesAvailability.create(vertx, client)
                 .onComplete(context.succeeding(pfa -> context.verify(() -> {
                     assertThat("Kubernetes version : " + pfa.getKubernetesVersion() + " is too old",
-                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.V1_16), CoreMatchers.is(not(lessThan(0))));
+                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.MINIMAL_SUPPORTED_VERSION), CoreMatchers.is(not(lessThan(0))));
                 })))
 
                 .compose(pfa -> {
@@ -172,7 +172,7 @@ public abstract class AbstractCustomResourceOperatorIT<C extends KubernetesClien
         PlatformFeaturesAvailability.create(vertx, client)
                 .onComplete(context.succeeding(pfa -> context.verify(() -> {
                     assertThat("Kubernetes version : " + pfa.getKubernetesVersion() + " is too old",
-                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.V1_16), CoreMatchers.is(not(lessThan(0))));
+                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.MINIMAL_SUPPORTED_VERSION), CoreMatchers.is(not(lessThan(0))));
                 })))
                 .compose(pfa -> {
                     LOGGER.info("Creating resource");
@@ -220,7 +220,7 @@ public abstract class AbstractCustomResourceOperatorIT<C extends KubernetesClien
         PlatformFeaturesAvailability.create(vertx, client)
                 .onComplete(context.succeeding(pfa -> context.verify(() -> {
                     assertThat("Kubernetes version : " + pfa.getKubernetesVersion() + " is too old",
-                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.V1_16), CoreMatchers.is(not(lessThan(0))));
+                            pfa.getKubernetesVersion().compareTo(KubernetesVersion.MINIMAL_SUPPORTED_VERSION), CoreMatchers.is(not(lessThan(0))));
                 })))
                 .compose(pfa -> {
                     LOGGER.info("Creating resource");